Test for Compare Function
Completed testing for the compare function. It looks like it works to me.  Also added a getScore funtion to get the compatability score. When calling this in the future the dog needs to be the caller due to setting the score to the profile at the end of the function